requirements.txt
discord.py
python-dotenv
pytz
requests
ffmpeg
PyNaCl
pip
youtube-dl
discord.py[voice]
openai
Pillow
stability-sdk